Re-class atlas and realtime to DRY and clean
