+++ src/main/java/at/hannibal2/skyhanni/utils/render/item/atlas/SkyHanniItemAtlas.kt
@@ -1,73 +1,38 @@
 package at.hannibal2.skyhanni.utils.render.item.atlas
 
 import at.hannibal2.skyhanni.test.command.ErrorManager
+import at.hannibal2.skyhanni.utils.render.item.SkyHanniAbstractItemTexture
 import at.hannibal2.skyhanni.utils.render.item.SkyHanniGuiItemRenderState
 import at.hannibal2.skyhanni.utils.render.item.SkyHanniItemRenderContext
-import com.mojang.blaze3d.ProjectionType
 import com.mojang.blaze3d.platform.TextureUtil
 import com.mojang.blaze3d.systems.RenderSystem
-import com.mojang.blaze3d.textures.FilterMode
 import com.mojang.blaze3d.textures.GpuTexture
-import com.mojang.blaze3d.textures.GpuTextureView
-import com.mojang.blaze3d.textures.TextureFormat
 import net.minecraft.client.Minecraft
-import net.minecraft.client.gui.render.TextureSetup
-import net.minecraft.client.gui.render.state.BlitRenderState
 import net.minecraft.client.gui.render.state.GuiRenderState
 import net.minecraft.client.renderer.CachedOrthoProjectionMatrixBuffer
-import net.minecraft.client.renderer.RenderPipelines
-import net.minecraft.client.renderer.texture.AbstractTexture
 import net.minecraft.client.renderer.texture.Dumpable
 import net.minecraft.resources.Identifier
 import java.nio.file.Path
 
-internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
+internal class SkyHanniItemAtlas : SkyHanniAbstractItemTexture(), Dumpable {
 
     companion object {
         private val identifier = Identifier.fromNamespaceAndPath("skyhanni", "item_atlas")
-        private const val PADDING = 1
     }
 
     init {
         Minecraft.getInstance().textureManager.register(identifier, this)
     }
 
-    private var depthTexture: GpuTexture? = null
-    private var depthTextureView: GpuTextureView? = null
-
     private var sizePixels = 0
-    private var root: Node? = null
+    private var packer: SkyHanniAtlasBinPacker? = null
+    private var renderer: SkyHanniItemAtlasRenderer? = null
     private var needsGrowing = false
     private val positions = HashMap<SkyHanniAtlasKey, SkyHanniItemAtlasEntry>()
 
-    /**
-     * Square-slot BSP bin packer, modelled after FontTexture.Node
-     */
-    private inner class Node(val x: Int, val y: Int, val width: Int, val height: Int) {
-        var left: Node? = null
-        var right: Node? = null
-        var occupied = false
-
-        fun insert(size: Int): Node? {
-            if (left != null || right != null) return left?.insert(size) ?: right?.insert(size)
-            if (occupied || size > width || size > height) return null
-            if (size == width && size == height) {
-                occupied = true
-                return this
-            }
-
-            val dw = width - size
-            val dh = height - size
-            if (dw > dh) {
-                left = Node(x, y, size, height)
-                right = Node(x + size + PADDING, y, width - size - PADDING, height)
-            } else {
-                left = Node(x, y, width, size)
-                right = Node(x, y + size + PADDING, width, height - size - PADDING)
-            }
-            return left?.insert(size)
-        }
-    }
+    val usage = GpuTexture.USAGE_RENDER_ATTACHMENT or
+        GpuTexture.USAGE_TEXTURE_BINDING or
+        GpuTexture.USAGE_COPY_SRC
 
     override fun dumpContents(id: Identifier, path: Path) {
         val texture = this.texture ?: return
@@ -82,6 +47,18 @@ internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
                 "path" to path.toString()
             )
         }
+    }
+
+    private fun ensureAllocated() {
+        if (texture != null) return
+        allocate(512.coerceAtMost(RenderSystem.getDevice().maxTextureSize))
+    }
+
+    private fun allocate(size: Int) {
+        sizePixels = size
+        allocateTextures(size, "SkyHanni item atlas", "SkyHanni item atlas depth", usage)
+        packer = SkyHanniAtlasBinPacker(size)
+        renderer = SkyHanniItemAtlasRenderer(size, textureView!!, depthTextureView!!, texture!!, depthTexture!!)
     }
 
     /**
@@ -109,29 +86,23 @@ internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
 
     private fun SkyHanniAnimatedAtlasKey.clearPreviousFrame() {
         val prevEntry = positions[this.copy(frameNumber = frameNumber - 1)] ?: return
-        clearSlot(prevEntry.x, prevEntry.y, prevEntry.pixelSize)
+        renderer?.clearSlot(prevEntry.x, prevEntry.y, prevEntry.pixelSize)
     }
 
     private fun recordPosition(key: SkyHanniAtlasKey, slotX: Int, slotY: Int, pixelSize: Int) {
         val u = slotX.toFloat() / sizePixels.toFloat()
         val v = (sizePixels - slotY).toFloat() / sizePixels.toFloat()
-        val entry = if (key is SkyHanniAnimatedAtlasKey) {
+        positions[key] = if (key is SkyHanniAnimatedAtlasKey) {
             SkyHanniAnimatedItemAtlasEntry(slotX, slotY, u, v, pixelSize, key.frameNumber)
         } else {
             SkyHanniItemAtlasEntry(slotX, slotY, u, v, pixelSize)
         }
-        positions[key] = entry
-    }
-
-    private fun ensureAllocated() {
-        if (texture != null) return
-        allocate(512.coerceAtMost(RenderSystem.getDevice().maxTextureSize))
     }
 
     private data class AtlasRenderJob(
         val key: SkyHanniAtlasKey,
         val representative: SkyHanniGuiItemRenderState,
-        val node: Node,
+        val node: SkyHanniAtlasBinPacker.PackedNode,
         val pixelSize: Int,
     )
 
@@ -142,6 +113,8 @@ internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
         pruneFrames(frameNumber)
         if (atlasStates.isEmpty()) return
         ensureAllocated()
+        val renderer = renderer ?: return
+        val packer = packer ?: return
 
         val groups = LinkedHashMap<SkyHanniAtlasKey, MutableList<SkyHanniGuiItemRenderState>>()
         for (state in atlasStates) groups.getOrPut(state.atlasKey) { mutableListOf() }.add(state)
@@ -158,7 +131,7 @@ internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
                 continue
             }
 
-            val node = root?.insert(neededPixels) ?: run {
+            val node = packer.insert(neededPixels) ?: run {
                 needsGrowing = true
                 // Overflow, submitBlitForState will return false and fall back to realtime
                 continue
@@ -170,78 +143,13 @@ internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
 
         if (renderJobs.isEmpty()) return
 
-        render(projectionBuffer) {
+        renderer.render(projectionBuffer) {
             for ((key, representative, node, pixelSize) in renderJobs) {
-                renderItemToAtlas(representative, node.x, node.y, pixelSize)
+                renderer.renderItemToAtlas(representative, node.x, node.y, pixelSize, bufferSource, featureRenderDispatcher)
                 recordPosition(key, node.x, node.y, pixelSize)
             }
             bufferSource.endBatch()
         }
-    }
-
-    fun invalidate() {
-        positions.clear()
-        root = null
-        close()
-    }
-
-    val usage = GpuTexture.USAGE_RENDER_ATTACHMENT or
-        GpuTexture.USAGE_TEXTURE_BINDING or
-        GpuTexture.USAGE_COPY_SRC
-
-    @Suppress("UnsafeCallOnNullableType")
-    private fun allocate(size: Int) {
-        sizePixels = size
-        root = Node(0, 0, size, size)
-        val device = RenderSystem.getDevice()
-        texture = device.createTexture("SkyHanni item atlas", usage, TextureFormat.RGBA8, size, size, 1, 1)
-            //? if < 1.21.11 {
-            .also { it.setTextureFilter(FilterMode.NEAREST, false) }
-        //?}
-        textureView = device.createTextureView(texture!!)
-        depthTexture = device.createTexture("SkyHanni item atlas depth", 8, TextureFormat.DEPTH32, size, size, 1, 1)
-        depthTextureView = device.createTextureView(depthTexture!!)
-        device.createCommandEncoder().clearColorAndDepthTextures(texture!!, 0, depthTexture!!, 1.0)
-    }
-
-    override fun close() {
-        textureView?.close()
-        textureView = null
-        texture?.close()
-        texture = null
-        depthTextureView?.close()
-        depthTextureView = null
-        depthTexture?.close()
-        depthTexture = null
-    }
-
-    private fun render(projectionBuffer: CachedOrthoProjectionMatrixBuffer, block: () -> Unit) {
-        val bufferSlice = projectionBuffer.getBuffer(sizePixels.toFloat(), sizePixels.toFloat())
-        RenderSystem.setProjectionMatrix(bufferSlice, ProjectionType.ORTHOGRAPHIC)
-        RenderSystem.outputColorTextureOverride = textureView
-        RenderSystem.outputDepthTextureOverride = depthTextureView
-        block()
-        RenderSystem.outputColorTextureOverride = null
-        RenderSystem.outputDepthTextureOverride = null
-        tryGrow()
-    }
-
-    private fun SkyHanniItemRenderContext.renderItemToAtlas(
-        shState: SkyHanniGuiItemRenderState,
-        slotX: Int,
-        slotY: Int,
-        pixelSize: Int,
-    ) {
-        RenderSystem.enableScissorForRenderTypeDraws(
-            slotX, sizePixels - slotY - pixelSize, pixelSize, pixelSize,
-        )
-        shState.renderItemToTexture(
-            bufferSource, featureRenderDispatcher,
-            centerX = slotX.toFloat() + pixelSize / 2.0f,
-            centerY = slotY.toFloat() + pixelSize / 2.0f,
-            pixelSize = pixelSize,
-        )
-        RenderSystem.disableScissorForRenderTypeDraws()
     }
 
     // Returns false if no atlas entry exists (overflow/not yet settled), caller falls back to realtime
@@ -256,44 +164,20 @@ internal class SkyHanniItemAtlas : AbstractTexture(), AutoCloseable, Dumpable {
                 entry.x, entry.y, entry.u, entry.v, entry.pixelSize, frameNumber
             )
         }
-        submitBlitRenderState(state, entry.u, entry.v, entry.pixelSize, guiRenderState)
+        renderer?.submitBlitForState(state, guiRenderState, entry)
         return true
     }
 
-    private fun submitBlitRenderState(
-        shState: SkyHanniGuiItemRenderState,
-        u: Float,
-        v: Float,
-        pixelSize: Int,
-        guiRenderState: GuiRenderState,
-    ) {
-        val textureView = this@SkyHanniItemAtlas.textureView ?: throw IllegalStateException("Atlas not allocated")
-        val size = sizePixels.toFloat()
-        val slotF = pixelSize.toFloat()
-        val u1 = u + slotF / size
-        val v1 = v + (-slotF) / size
-        guiRenderState.submitBlitToCurrentLayer(
-            BlitRenderState(
-                RenderPipelines.GUI_TEXTURED,
-                //? if < 1.21.11 {
-                TextureSetup.singleTexture(textureView),
-                //?} else
-                // TextureSetup.singleTexture(textureView, RenderSystem.getSamplerCache().getRepeat(FilterMode.NEAREST)),
-                shState.pose(),
-                shState.x0(), shState.y0(), shState.x1(), shState.y1(),
-                u, u1, v, v1,
-                -1,
-                shState.scissorArea(),
-            )
-        )
+    fun invalidate() {
+        positions.clear()
+        packer = null
+        renderer = null
+        close()
     }
 
-    private fun clearSlot(x: Int, y: Int, size: Int) {
-        val texture = this.texture ?: throw IllegalStateException("Atlas not allocated")
-        val depthTexture = this.depthTexture ?: throw IllegalStateException("Atlas not allocated")
-        RenderSystem.getDevice().createCommandEncoder().clearColorAndDepthTextures(
-            texture, 0, depthTexture, 1.0,
-            x, sizePixels - y - size, size, size,
-        )
+    override fun close() {
+        super.close()
+        packer = null
+        renderer = null
     }
 }
